Initial type model for Signature
This is very ugly and Stringly Typed at the moment.  It will become more
elegant as I learn the types that fit well with existing Haskell HTTP
and web frameworks like Happstack, Snap, and Yesod.

There are clearly no useful behaviors yet and no notion of what this
will depend on in practice.